split entrypoint for modules bootloader
